Fixes/updates
Timeout motor after 30 secs
Correct types for battery monitoring
Remove devboard IO
Permit tasks to be restarted after exit

+++ src/trv-state.h
@@ -12,7 +12,7 @@ typedef struct trv_state_s
 {
   // Read only
   float temperature;
-  uint16_t batteryRaw;
+  uint32_t batteryRaw;
   uint8_t batteryPercent;
   uint8_t isCharging;
   uint8_t valve_position;
